Refactor server entry point and update test imports: change app reference to server in test files and modify server listening logic for better module handling
