+++ backend/tests/routes/cloudinaryRoutes.test.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const request = require('supertest');
-const app = require('../../app'); // Your app's entry point
+const app = require('../../server'); // Your app's entry point
 const mongoose = require('mongoose');
 const { cleanupTask } = require('../../tasks/cleanupTokens'); // Import the cleanup task
